+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Infrastructure
+* Fix gathering licenses of typeshed libraries
+* Fix spellcheck erroring out on LICENSE file
+
 ## [3.10.1] - 2023-09-27
 
 ### Fixed